Sort functionality complete
